// components/home.types.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = {
  [K in keyof T]: T[K];
};
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & {
  [SubKey in K]?: Maybe<T[SubKey]>;
};
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & {
  [SubKey in K]: Maybe<T[SubKey]>;
};
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  BlockchainData: any;
  BroadcastId: any;
  ChainId: any;
  CollectModuleData: any;
  ContractAddress: any;
  CreateHandle: any;
  Cursor: any;
  DateTime: any;
  Ens: any;
  EthereumAddress: any;
  FollowModuleData: any;
  Handle: any;
  HandleClaimIdScalar: any;
  InternalPublicationId: any;
  Jwt: any;
  LimitScalar: any;
  Locale: any;
  Markdown: any;
  MimeType: any;
  NftOwnershipId: any;
  Nonce: any;
  NotificationId: any;
  ProfileId: any;
  PublicationId: any;
  PublicationTag: any;
  PublicationUrl: any;
  ReferenceModuleData: any;
  Search: any;
  Signature: any;
  Sources: any;
  TimestampScalar: any;
  TxHash: any;
  TxId: any;
  UnixTimestamp: any;
  Url: any;
  Void: any;
};
// ...
export type Attribute = {
  __typename?: "Attribute";
  /** The display type */
  displayType?: Maybe<Scalars["String"]>;
  /** identifier of this attribute, we will update by this id  */
  key: Scalars["String"];
  /** The trait type - can be anything its the name it will render so include spaces */
  traitType?: Maybe<Scalars["String"]>;
  /** Value attribute */
  value: Scalars["String"];
};
// ...
export type Dispatcher = {
  __typename?: "Dispatcher";
  /** The dispatcher address */
  address: Scalars["EthereumAddress"];
  /** If the dispatcher can use the relay */
  canUseRelay: Scalars["Boolean"];
};
// ...
export type EnsOnChainIdentity = {
  __typename?: "EnsOnChainIdentity";
  /** The default ens mapped to this address */
  name?: Maybe<Scalars["Ens"]>;
};

/** The erc20 type */
export type Erc20 = {
  __typename?: "Erc20";
  /** The erc20 address */
  address: Scalars["ContractAddress"];
  /** Decimal places for the token */
  decimals: Scalars["Int"];
  /** Name of the symbol */
  name: Scalars["String"];
  /** Symbol for the token */
  symbol: Scalars["String"];
};
// ...
export type FeeFollowModuleSettings = {
  __typename?: "FeeFollowModuleSettings";
  /** The collect module amount info */
  amount: ModuleFeeAmount;
  contractAddress: Scalars["ContractAddress"];
  /** The collect module recipient address */
  recipient: Scalars["EthereumAddress"];
  /** The follow modules enum */
  type: FollowModules;
};

export type Follow = {
  followModule?: InputMaybe<FollowModuleRedeemParams>;
  profile: Scalars["ProfileId"];
};

export type FollowModule =
  | FeeFollowModuleSettings
  | ProfileFollowModuleSettings
  | RevertFollowModuleSettings;
// ...
export type Media = {
  __typename?: "Media";
  /** Height - will always be null on the public API */
  height?: Maybe<Scalars["Int"]>;
  /** The image/audio/video mime type for the publication */
  mimeType?: Maybe<Scalars["MimeType"]>;
  /** Size - will always be null on the public API */
  size?: Maybe<Scalars["Int"]>;
  /** The token image nft */
  url: Scalars["Url"];
  /** Width - will always be null on the public API */
  width?: Maybe<Scalars["Int"]>;
};
// ...
export type NftImage = {
  __typename?: "NftImage";
  /** The token image nft */
  chainId: Scalars["Int"];
  /** The contract address */
  contractAddress: Scalars["ContractAddress"];
  /** The token id of the nft */
  tokenId: Scalars["String"];
  /** The token image nft */
  uri: Scalars["Url"];
  /** If the NFT is verified */
  verified: Scalars["Boolean"];
};
// ...
export type OnChainIdentity = {
  __typename?: "OnChainIdentity";
  /** The ens information */
  ens?: Maybe<EnsOnChainIdentity>;
  /** The POH status */
  proofOfHumanity: Scalars["Boolean"];
  /** The sybil dot org information */
  sybilDotOrg: SybilDotOrgIdentity;
};
// ...
export type Profile = {
  __typename?: "Profile";
  /** Optionals param to add extra attributes on the metadata */
  attributes?: Maybe<Array<Attribute>>;
  /** Bio of the profile */
  bio?: Maybe<Scalars["String"]>;
  /** The cover picture for the profile */
  coverPicture?: Maybe<ProfileMedia>;
  /** The dispatcher */
  dispatcher?: Maybe<Dispatcher>;
  /** The follow module */
  followModule?: Maybe<FollowModule>;
  /** Follow nft address */
  followNftAddress?: Maybe<Scalars["ContractAddress"]>;
  /** The profile handle */
  handle?: Scalars["Handle"];
  /** The profile id */
  id?: Scalars["ProfileId"];
  /** Is the profile default */
  isDefault?: Scalars["Boolean"];
  isFollowedByMe?: Scalars["Boolean"];
  isFollowing?: Scalars["Boolean"];
  /** Metadata url */
  metadata?: Maybe<Scalars["Url"]>;
  /** Name of the profile */
  name?: Maybe<Scalars["String"]>;
  /** The on chain identity */
  onChainIdentity?: OnChainIdentity;
  /** Who owns the profile */
  ownedBy?: Scalars["EthereumAddress"];
  /** The picture for the profile */
  picture?: Maybe<ProfileMedia>;
  /** Profile stats */
  stats?: ProfileStats;
};

/** The Profile */
export type ProfileIsFollowingArgs = {
  who?: InputMaybe<Scalars["ProfileId"]>;
};

export type ProfileFollowModuleBeenRedeemedRequest = {
  followProfileId: Scalars["ProfileId"];
  redeemingProfileId: Scalars["ProfileId"];
};

export type ProfileFollowModuleRedeemParams = {
  /** The profile id to use to follow this profile */
  profileId: Scalars["ProfileId"];
};

export type ProfileFollowModuleSettings = {
  __typename?: "ProfileFollowModuleSettings";
  contractAddress: Scalars["ContractAddress"];
  /** The follow module enum */
  type: FollowModules;
};

export type ProfileFollowRevenueQueryRequest = {
  /** The profile id */
  profileId: Scalars["ProfileId"];
};

export type ProfileMedia = Media | NftImage;
// ...
export type ProfileStats = {
  __typename?: "ProfileStats";
  commentsTotal: Scalars["Int"];
  id: Scalars["ProfileId"];
  mirrorsTotal: Scalars["Int"];
  postsTotal: Scalars["Int"];
  publicationsTotal: Scalars["Int"];
  /** Total collects count */
  totalCollects: Scalars["Int"];
  /** Total comment count */
  totalComments: Scalars["Int"];
  /** Total follower count */
  totalFollowers: Scalars["Int"];
  /** Total following count (remember the wallet follows not profile so will be same for every profile they own) */
  totalFollowing: Scalars["Int"];
  /** Total mirror count */
  totalMirrors: Scalars["Int"];
  /** Total post count */
  totalPosts: Scalars["Int"];
  /** Total publication count */
  totalPublications: Scalars["Int"];
};
// ...
export type RevertFollowModuleSettings = {
  __typename?: "RevertFollowModuleSettings";
  contractAddress: Scalars["ContractAddress"];
  /** The follow module enum */
  type: FollowModules;
};
// ...
export type SybilDotOrgIdentity = {
  __typename?: "SybilDotOrgIdentity";
  source: SybilDotOrgIdentitySource;
  /** The sybil dot org status */
  verified: Scalars["Boolean"];
};
// ...
export enum MediaType {
  Video,
  Image,
  Gif,
}

export interface UploadedMedia {
  cid: string;
  type: MediaType;
}